refactoring on ortho- and perspective-viewports related files, exports and docs
