+++ src/core/viewports/orthographic-viewport.js
@@ -18,33 +18,33 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-import Viewport from '../../viewports/viewport';
+import Viewport from './viewport';
 import mat4_lookAt from 'gl-mat4/lookAt';
-import mat4_perspective from 'gl-mat4/perspective';
+import mat4_ortho from 'gl-mat4/ortho';
 
-const DEGREES_TO_RADIANS = Math.PI / 180;
-
-export default class PerspectiveViewport extends Viewport {
+export default class OrthographicViewport extends Viewport {
   constructor({
     // viewport arguments
     width, // Width of viewport
     height, // Height of viewport
     // view matrix arguments
-    eye, // Defines eye position
+    eye = [0, 0, 1], // Defines eye position, default unit distance along z axis
     lookAt = [0, 0, 0], // Which point is camera looking at, default origin
     up = [0, 1, 0], // Defines up direction, default positive y axis
     // projection matrix arguments
-    fovy = 75, // Field of view covered by camera
     near = 1, // Distance of near clipping plane
     far = 100, // Distance of far clipping plane
+    left, // Left bound of the frustum
+    top, // Top bound of the frustum
     // automatically calculated
-    aspect = null // Aspect ratio (set to viewport widht/height)
+    right = null, // Right bound of the frustum
+    bottom = null // Bottom bound of the frustum
   }) {
-    const fovyRadians = fovy * DEGREES_TO_RADIANS;
-    aspect = Number.isFinite(aspect) ? aspect : width / height;
+    right = Number.isFinite(right) ? right : left + width;
+    bottom = Number.isFinite(bottom) ? bottom : top + height;
     super({
       viewMatrix: mat4_lookAt([], eye, lookAt, up),
-      projectionMatrix: mat4_perspective([], fovyRadians, aspect, near, far),
+      projectionMatrix: mat4_ortho([], left, right, bottom, top, near, far),
       width,
       height
     });
